requirements.txt
datasets
pandas
sqlalchemy
fastapi
uvicorn
requests
python-dotenv
streamlit
